Change how we kill personal condors. Use condor_off -master first on all platforms. If we do start killing daemons, kill master last. Call it quits on waiting after a couple times through checking our pid list. #3594

+++ src/condor_scripts/CondorPersonal.pm
@@ -344,6 +344,7 @@ sub StartCondorWithParams
 	CondorPersonal::Reset();
 	debug( "StartCondor config_and_port is --$config_and_port--\n",$debuglevel);
 	debug( "Personal Condor Started\n",$debuglevel);
+	system("date");
 	print  "Personal Condor Started\n";
 	print scalar localtime() . "\n";
 	return( $config_and_port );
@@ -369,6 +370,9 @@ sub debug {
 sub debug_flush {
 	print "\nDEBUG_FLUSH:\n";
 	my $logdir = `condor_config_val log`;
+	$_ = $logdir;
+	s/\\/\//g;
+	$logdir = $_;
 	fullchomp($logdir);
 	print "\nLog directory: $logdir and contains:\n";
 	system("ls -lh $logdir");
@@ -1328,6 +1332,7 @@ sub StartPersonalCondor
 	}
 
 	my $res = IsRunningYet();
+	system("date");
 	print "Back from IsRunningYet in start personal condor\n";
 	if($res == 0) {
 		debug_flush();
@@ -1931,10 +1936,14 @@ sub KillDaemonPids
 {
 	my $desiredconfig = shift;
 	my $oldconfig = $ENV{CONDOR_CONFIG};
+	print "KillDaemonPids: Current: $oldconfig Killing: $desiredconfig\n";
 	$ENV{CONDOR_CONFIG} = $desiredconfig;
 	my $logdir = `condor_config_val log`;
 	$logdir =~ s/\012+$//;
 	$logdir =~ s/\015+$//;
+	$_ = $logdir;
+	s/\\/\//g;
+	$logdir = $_;
 	my $masterpid = 0;
 	my $cnt = 0;
 	my $cmd;
@@ -1943,6 +1952,8 @@ sub KillDaemonPids
 
 	print "Before the Kill of $desiredconfig\n";
 	CollectWhoPids($logdir);
+
+	CondorTest::runToolNTimes("condor_off -master",1,0);
 
 	if($isnightly) {
 		DisplayPartialLocalConfig($desiredconfig);
@@ -1954,25 +1965,25 @@ sub KillDaemonPids
 	debug("Asked to kill: $oldconfig\n",$debuglevel);
 	my $thispid = 0;
 	# first find the master and use a kill 3(fast kill)
-	open(PD,"<$pidfile") or die "Can not open :$pidfile:$!\n";
-	while(<PD>) {
-		fullchomp();
-		$thispid = $_;
-		print "$thispid\n";
-		if($thispid =~ /^(\d+)\s+MASTER.*$/) {
-			print "fast kill on master $1\n";
-			$masterpid = $1;
-			if(CondorUtils::is_windows() == 1) {
-				$cmd = "taskkill /PID $masterpid /T /F";
-				system($cmd);
-			} else {
-				$cnt = kill 3, $masterpid;
-			}
-			debug("Gentle kill for master: $masterpid $thispid($cnt)\n",$debuglevel);
-			last;
-		}
-	}
-	close(PD);
+	#open(PD,"<$pidfile") or die "Can not open :$pidfile:$!\n";
+	#while(<PD>) {
+		#fullchomp();
+		#$thispid = $_;
+		#print "$thispid\n";
+		#if($thispid =~ /^(\d+)\s+MASTER.*$/) {
+			#print "fast kill on master $1\n";
+			#$masterpid = $1;
+			#if(CondorUtils::is_windows() == 1) {
+				#$cmd = "taskkill /PID $masterpid /T /F";
+				#system($cmd);
+			#} else {
+				#$cnt = kill 3, $masterpid;
+			#}
+			#debug("Gentle kill for master: $masterpid $thispid($cnt)\n",$debuglevel);
+			#last;
+		#}
+	#}
+	#close(PD);
 	# give it a little time for a shutdown
 	sleep(10);
 	my $res = CheckPids($pidfile);
@@ -2004,13 +2015,14 @@ sub KillDaemonPids
 	# did it work.... is process still around? after 3 tries, where are we?
 	print "Last result was: $res\n";
 	if($res eq "going away") {
-		while($res ne "all gone") {
+		#while($res ne "all gone") {
 			sleep(10);
 			$res = CheckPids($pidfile);
-		}
+		#}
 	} elsif(($res eq "stubborn") || ($res eq "master gone")) {
 			sleep(10);
 			$res = CheckPids($pidfile,"kill all");
+			$res = CheckPids($pidfile,"kill master");
 			print "imposed sudden death, where are we now?\n";
 			sleep(10);
 			$res = CheckPids($pidfile,);
@@ -2046,29 +2058,59 @@ sub CheckPids
 	my $masterlives = 0;
 	my $otherslive = 0;
 	my $howmanydeamons = 0;
+	
+	system("date");
+	if(defined $action) {
+		print "CheckPids: $pidfile Action: $action\n";
+	} else {
+		print "CheckPids: $pidfile Action: none\n";
+	}
 	open(PF, "<$pidfile") or die "Failed to find pid file:$pidfile:$1\n";
 	while(<PF>) {
 		chomp();
 		$line = $_;
+		my $daemon = "";
+		my $pid = "";
 		if($line =~ /(\d+)\s+(\w+).*/) {
+			$daemon = $2;
+			$pid = $1;
 			$howmanydeamons += 1;
 			@grift = ();
 			if(CondorUtils::is_windows() == 1) {
-				@grift = `tasklist | grep $1`;
+				@grift = `tasklist | grep $pid`;
+				my $realpid = "";
 				$linecount = @grift;
 				if($linecount == 0) {
-					print "Daemon $2 PID $1 is gone\n";
-				} elsif($linecount == 1) {
-					print "Daemon $2 PID $1 is still alive\n";
-					if($2 eq "MASTER") {
-						$masterlives += 1;
-					} else {
-						$otherslive += 1;
+					print "Daemon $daemon PID $pid is gone\n";
+				} else {
+					print "Fishing for status: $daemon/$pid\n";
+					foreach my $ent (@grift){
+						chomp($ent);
+						if($ent =~ /(.*?)\s+(\d+)\s+(\w+).*/) {
+							if($2 eq $pid) {
+								print "Daemon $daemon PID $pid is still alive\n";
+								if($daemon eq "MASTER") {
+									$masterlives += 1;
+								} else {
+									$otherslive += 1;
+								}
+							} else {
+								print "Ignoring $1 pid $2\n";
+							}
+						}
 					}
 					if(defined $action) {
 						if($action eq "kill all") {
-							my $cmd = "taskkill /PID $1 /F";
-							system($cmd);
+							if($daemon ne "MASTER") {
+								my $cmd = "taskkill /PID $pid /T /F";
+								system($cmd);
+							}
+						}
+						if($action eq "kill master") {
+							if($daemon eq "MASTER") {
+								my $cmd = "taskkill /PID $pid /T /F";
+								system($cmd);
+							}
 						}
 					}
 				}
@@ -2086,7 +2128,14 @@ sub CheckPids
 					}
 					if(defined $action) {
 						if($action eq "kill all") {
-							kill 15, $1;
+							if($daemon ne "MASTER") {
+								kill 15, $1;
+							}
+						}
+						if($action eq "kill master") {
+							if($daemon eq "MASTER") {
+								kill 15, $1;
+							}
 						}
 					}
 				} else {
@@ -2320,6 +2369,9 @@ sub DisplayPartialLocalConfig
 {
 	my $configloc = shift;
 	my $logdir = `condor_config_val log`;
+	$_ = $logdir;
+	s/\\/\//g;
+	$logdir = $_;
 	my $fullpathtolocalconfig = "";
 	my $line = "";
 	fullchomp($logdir);
